requirements.txt
Django>=5.0,<5.1
gunicorn
python-decouple
psycopg[binary]
dj-database-url
requests
djangorestframework